[Part 4] Add answers to sections c and d.

+++ part4/blogilista/controllers/blogs.js
@@ -1,8 +1,11 @@
 const blogsRouter = require('express').Router()
+const middleware = require('../utils/middleware')
 const Blog = require('../models/blog')
 
 blogsRouter.get('/', async (_, response) => {
-  const blogs = await Blog.find({})
+  const blogs = await Blog
+    .find({}).populate('user', { username: 1, name: 1 })
+
   response.json(blogs)
 })
 
@@ -11,9 +14,22 @@ blogsRouter.get('/:id', async (request, response) => {
   blog ? response.json(blog) : response.status(404).end()
 })
 
-blogsRouter.post('/', async (request, response) => {
-  const blog = new Blog(request.body)
+blogsRouter.post('/', middleware.userExtractor, async (request, response) => {
+  const body = request.body
+  const user = request.user
+
+  const blog = new Blog({
+    title: body.title,
+    author: body.author,
+    url: body.url,
+    likes: body.likes ?? 0,
+    user: user._id
+  })
+
   const savedBlog = await blog.save()
+  user.blogs = user.blogs.concat(savedBlog._id)
+  await user.save()
+
   response.status(201).json(savedBlog)
 })
 
@@ -29,16 +45,26 @@ blogsRouter.put('/:id', async (request, response) => {
     title: request.body.title ?? match.title,
     author: request.body.author ?? match.author,
     url: request.body.url ?? match.url,
-    likes: request.body.likes ?? match.likes
+    likes: request.body.likes ?? match.likes ?? 0
   }
 
   const updatedBlog = await Blog.findByIdAndUpdate(id, blogToUpdate)
   response.status(200).json(updatedBlog)
 })
 
-blogsRouter.delete('/:id', async (request, response) => {
-  await Blog.findByIdAndDelete(request.params.id)
-  response.status(204).end()
+blogsRouter.delete('/:id', middleware.userExtractor, async (request, response) => {
+  const user = request.user
+
+  const blog = await Blog.findById(request.params.id)
+
+  if (!blog) {
+    response.status(404).end()
+  }
+
+  if (blog.user.toString() === user.id.toString()) {
+    await Blog.findByIdAndDelete(request.params.id)
+    response.status(204).end()
+  }
 })
 
 module.exports = blogsRouter
